providers/aws: Convert AWS Security Group to upstream aws-sdk-go
Moves structure test lib too, to that `make test` actually runs it

+++ builtin/providers/aws/structure_sdk_test.go
@@ -434,7 +434,7 @@ func TestFlattenAttachmentSDK(t *testing.T) {
 		t.Fatalf("expected instance to be i-00001, but got %s", result["instance"])
 	}
 
-	if result["device_index"] != 1 {
+	if result["device_index"] != int64(1) {
 		t.Fatalf("expected device_index to be 1, but got %d", result["device_index"])
 	}
